Feat : 임시 비밀번호 발급 기능 구현

+++ app/src/main/java/com/hana/fieldmate/data/remote/model/response/BusinessRes.kt
@@ -39,7 +39,7 @@ data class BusinessRes(
     @SerializedName("businessPeriodDto")
     val businessPeriod: BusinessPeriod,
     @SerializedName("revenue")
-    val revenue: Int,
+    val revenue: Long,
     @SerializedName("memberDtoList")
     val memberDtoList: List<MemberNameRes>,
     @SerializedName("description")
